c
A player begins with six, fair, six-sided dice, and has the goal of rolling them all and getting exactly one
outcome of each type {1, 2, 3, 4, 5, 6}. The player is allowed to roll the dice up to 3 times, and after each roll
they may set aside as many dice as they wish, and re-roll the remaining dice. Since the dice are fair, you may
convince yourself that the best strategy is to set aside those dice corresponding to numbers that have not yet
been rolled, and re-roll all duplicates.

Write a simulation to estimate the probability the player is successful in 3 or fewer rolls if they follow
the optimal strategy. The simulation should be based on at least 100,000 replications.